split

+++ init.sh
@@ -1,29 +1,57 @@
-# Disable Gatekeeper
-sudo spctl --master-disable
+exists()
+{
+  command -v "$1" >/dev/null 2>&1
+}
+
+# brew
+if ! exists brew; then
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+fi
+
+if ! exists brew command command-not-found-init; then
+  brew tap homebrew/command-not-found
+fi
 
 # base
-brew install curl wget
+if ! exists curl; then
+  brew install curl
+fi
+if ! exists wget; then
+  brew install wget
+fi
 
-# install nvm
-curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.2/install.sh | bash
-# install oh-my-zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-# install oh-my-zsh theme powerlevel9k
-git clone https://github.com/bhilburn/powerlevel9k.git ~/.oh-my-zsh/custom/themes/powerlevel9k
-# install oh-my-zsh plugin zsh-autosuggestions
-git clone git://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-# install oh-my-zsh plugin zsh-syntax-highlighting
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-# make symbolic links
-ln -s ~/dotfiles/.zshrc ~/.zshrc
+# oh-my-zsh
+if [ ! -d ~/.oh-my-zsh ]; then
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+  # theme powerlevel9k
+  git clone https://github.com/bhilburn/powerlevel9k.git ~/.oh-my-zsh/custom/themes/powerlevel9k
+  # plugin zsh-autosuggestions
+  git clone git://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+  # plugin zsh-syntax-highlighting
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
+  # symbolic link
+  ln -s ~/dotfiles/.zshrc ~/.zshrc
+fi
 
-# command-not-found
-brew tap homebrew/command-not-found
 # node
-nvm install node
-npm i -g http-server
-# brew
-brew install yarn cloc autojump
-# cask
-brew tap caskroom/cask
-brew cask install 1password mpv google-chrome visual-studio-code iterm2 sourcetree gas-mask dash diffmerge smcfancontrol xee the-unarchiver wireshark synergy cheatsheet xmind neteasemusic adobe-photoshop-cc sketch microsoft-office parallels-desktop
+if [ ! -d ~/.nvm ]; then
+  curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.2/install.sh | bash
+fi
+
+if ! exists node; then
+  nvm install node
+fi
+
+# utils
+if ! exists http-server; then
+  npm i -g http-server
+fi
+if ! exists yarn; then
+  brew install yarn
+fi
+if ! exists cloc; then
+  brew install cloc
+fi
+if ! exists autojump; then
+  brew install autojump
+fi
